feat : used eslint AirBnB style and increase test coverage

+++ src/Applications/use_case/ReplyUseCase.js
@@ -1,4 +1,7 @@
-const AddReply = require("../../Domains/replies/entities/AddReply");
+/* eslint-disable no-underscore-dangle */
+const AddReply = require('../../Domains/replies/entities/AddReply');
+const NotFoundError = require('../../Commons/exceptions/NotFoundError');
+const AuthorizationError = require('../../Commons/exceptions/AuthorizationError');
 
 class ReplyUseCase {
   constructor({ threadRepository, commentRepository, replyRepository }) {
@@ -9,17 +12,14 @@ class ReplyUseCase {
 
   async addReply(useCasePayload, threadId, commentId, owner) {
     const threadCount = await this._threadRepository.verifyThreadAvailability(
-      threadId
+      threadId,
     );
     if (!threadCount) {
-      const NotFoundError = require("../../Commons/exceptions/NotFoundError");
-      throw new NotFoundError("thread tidak ditemukan");
+      throw new NotFoundError('thread tidak ditemukan');
     }
-    const commentCount =
-      await this._commentRepository.verifyCommentAvailability(commentId);
+    const commentCount = await this._commentRepository.verifyCommentAvailability(commentId);
     if (!commentCount) {
-      const NotFoundError = require("../../Commons/exceptions/NotFoundError");
-      throw new NotFoundError("komentar tidak ditemukan");
+      throw new NotFoundError('komentar tidak ditemukan');
     }
     const addReply = new AddReply({ ...useCasePayload, commentId, owner });
     return this._replyRepository.addReply(addReply);
@@ -27,26 +27,25 @@ class ReplyUseCase {
 
   async deleteReply(threadId, commentId, replyId, owner) {
     const threadCount = await this._threadRepository.verifyThreadAvailability(
-      threadId
+      threadId,
     );
     if (!threadCount) {
-      const NotFoundError = require("../../Commons/exceptions/NotFoundError");
-      throw new NotFoundError("thread tidak ditemukan");
+      throw new NotFoundError('thread tidak ditemukan');
     }
-    const commentCount =
-      await this._commentRepository.verifyCommentAvailability(commentId);
+    const commentCount = await this._commentRepository.verifyCommentAvailability(commentId);
     if (!commentCount) {
-      const NotFoundError = require("../../Commons/exceptions/NotFoundError");
-      throw new NotFoundError("komentar tidak ditemukan");
+      throw new NotFoundError('komentar tidak ditemukan');
     }
-    const reply = await this._replyRepository.verifyReplyOwner(replyId, owner);
+    const replyExists = await this._replyRepository.verifyReplyAvailability(replyId);
+    if (!replyExists) {
+      throw new NotFoundError('balasan tidak ditemukan');
+    }
+    const reply = await this._replyRepository.verifyReplyOwner(replyId);
     if (!reply) {
-      const NotFoundError = require("../../Commons/exceptions/NotFoundError");
-      throw new NotFoundError("balasan tidak ditemukan");
+      throw new NotFoundError('balasan tidak ditemukan');
     }
     if (reply.owner !== owner) {
-      const AuthorizationError = require("../../Commons/exceptions/AuthorizationError");
-      throw new AuthorizationError("anda tidak berhak mengakses resource ini");
+      throw new AuthorizationError('anda tidak berhak mengakses resource ini');
     }
     await this._replyRepository.deleteReply(replyId);
   }
